backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.1.0
Pillow==10.4.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.0
redis==5.0.1
django-otp==1.5.0
pyotp==2.9.0
qrcode[pil]==7.4.2
